+++ Assets/Script/StartScene/StartButtonManager.cs
@@ -22,7 +22,8 @@ public class StartButtonManager : MonoBehaviour
     public void OnClickNewStart()
     {
         File.Delete(Application.dataPath + "/SaveFile.dat");
-        File.Delete(Application.dataPath + "/Save/InventoryData.xml");
+        if (File.Exists(Application.dataPath + "/Save/InventoryData.xml"))
+            File.Delete(Application.dataPath + "/Save/InventoryData.xml");
         SceneManager.LoadScene("Main");
 
     }
